# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Seattle Solvers Tools API",
    "description": "FTC parts inventory, search, and receipt reading for tools.seattlesolvers.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/receipts/upload": {
      "post": {
        "summary": "Upload a receipt file — a digital PDF invoice, or a photo of a paper receipt",
        "tags": [
          "receipts"
        ],
        "description": "multipart/form-data with fields `vendor` and `file`. A PDF is read from its text layer, so a downloaded invoice needs no OCR. An image (jpg/png/webp/gif) is OCR'd, falling back to Claude vision only when that is not good enough — this is the physical-receipt path. For an emailed confirmation prefer POST /receipts, which is exact and cheaper.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "vendor": {
                      "type": "string",
                      "enum": [
                        "GOBILDA",
                        "REV",
                        "AXON",
                        "FERRA",
                        "MELONBOTICS",
                        "OFFSET",
                        "MATA",
                        "UXCELL",
                        "OTHER"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PROCESSING",
                        "PARSED",
                        "CONFIRMED",
                        "FAILED"
                      ]
                    },
                    "method": {
                      "type": "string",
                      "enum": [
                        "PASTED_TEXT",
                        "PASTED_HTML",
                        "PDF_TEXT",
                        "TESSERACT",
                        "CLAUDE_TEXT",
                        "CLAUDE_VISION"
                      ],
                      "nullable": true
                    },
                    "hasFile": {
                      "type": "boolean"
                    },
                    "orderTotal": {
                      "type": "number",
                      "nullable": true
                    },
                    "purchasedAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "lineItems": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "rawText": {
                            "type": "string",
                            "nullable": true
                          },
                          "parsedName": {
                            "type": "string"
                          },
                          "parsedSku": {
                            "type": "string",
                            "nullable": true
                          },
                          "quantity": {
                            "type": "integer"
                          },
                          "unitPrice": {
                            "type": "number",
                            "nullable": true
                          },
                          "lineTotal": {
                            "type": "number",
                            "nullable": true
                          },
                          "matchConfidence": {
                            "type": "number"
                          },
                          "applied": {
                            "type": "boolean"
                          },
                          "matchedPart": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "sku": {
                                "type": "string",
                                "nullable": true
                              },
                              "manufacturer": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "name",
                              "sku",
                              "manufacturer"
                            ],
                            "additionalProperties": false,
                            "nullable": true
                          }
                        },
                        "required": [
                          "id",
                          "rawText",
                          "parsedName",
                          "parsedSku",
                          "quantity",
                          "unitPrice",
                          "lineTotal",
                          "matchConfidence",
                          "applied",
                          "matchedPart"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "id",
                    "vendor",
                    "status",
                    "method",
                    "hasFile",
                    "orderTotal",
                    "purchasedAt",
                    "createdAt",
                    "lineItems"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/receipts/{id}/file": {
      "get": {
        "summary": "Get a short-lived link to a receipt's original file",
        "tags": [
          "receipts"
        ],
        "description": "The receipts bucket is private, because order confirmations routinely carry a name and shipping address. There is therefore no durable URL to embed: call this when `hasFile` is true and use the returned link, which expires. Returns JSON rather than a redirect so the URL can go straight into an <img> or <a>, which cannot send an Authorization header.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "url",
                    "expiresIn"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/receipts/{id}/lines/{lineId}": {
      "patch": {
        "summary": "Correct a parsed line item (fix the matched part, quantity, or name)",
        "tags": [
          "receipts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "matchedPartId": {
                    "type": "string",
                    "nullable": true
                  },
                  "quantity": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "parsedName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "parsedSku": {
                    "type": "string",
                    "maxLength": 120,
                    "nullable": true
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "lineId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "rawText": {
                      "type": "string",
                      "nullable": true
                    },
                    "parsedName": {
                      "type": "string"
                    },
                    "parsedSku": {
                      "type": "string",
                      "nullable": true
                    },
                    "quantity": {
                      "type": "integer"
                    },
                    "unitPrice": {
                      "type": "number",
                      "nullable": true
                    },
                    "lineTotal": {
                      "type": "number",
                      "nullable": true
                    },
                    "matchConfidence": {
                      "type": "number"
                    },
                    "applied": {
                      "type": "boolean"
                    },
                    "matchedPart": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "sku": {
                          "type": "string",
                          "nullable": true
                        },
                        "manufacturer": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "sku",
                        "manufacturer"
                      ],
                      "additionalProperties": false,
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "rawText",
                    "parsedName",
                    "parsedSku",
                    "quantity",
                    "unitPrice",
                    "lineTotal",
                    "matchConfidence",
                    "applied",
                    "matchedPart"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}